Add creation of datasets and json data, implement line chart using the json data

+++ La-Game/La-Game/Models/Chart.cs
@@ -15,6 +15,6 @@ namespace La_Game.Models
         public string hoverBackgroundColor { get; set; }
         public string borderWidth { get; set; }
         public bool fill { get; set; }
-        public int[] data { get; set; }
+        public double[] data { get; set; }
     }
 }
